refactor: split conductor.rs into mod/helpers/tests modules
conductor.rs (1878 lines) was too large for single-context output.
Split into three focused files:
- conductor/mod.rs: TuiConductor struct and all impl methods (~500 lines)
- conductor/helpers.rs: free utility functions (field builders, form readers, record mappers)
- conductor/tests.rs: all 30 tests extracted to dedicated module

All 108 tests pass. No behaviour changed.
